test: add vote-on-poll test

+++ src/domain/application/use-cases/vote-on-poll.ts
@@ -1,5 +1,8 @@
 import { Vote } from "../../enterprise/entities/vote";
 import { UniqueEntityId } from "../../enterprise/object-values/unique-entity-id";
+import { DuplicateVoteError } from "../errors/duplicate-vote-error";
+import { InvalidPollOptionError } from "../errors/invalid-poll-option-error";
+import { PollNotFoundError } from "../errors/poll-not-found-error";
 import { PollRepository } from "../repositories/poll-repository";
 
 interface VoteOnPollUseCaseRequest {
@@ -20,17 +23,31 @@ export class VoteOnPollUseCase {
     pollOptionId,
     userId,
   }: VoteOnPollUseCaseRequest): Promise<VoteOnPollUseCaseResponse> {
+    const pollWithOptions = await this.pollRepository.findPollByIdWithOptions(
+      pollId
+    );
+    if (!pollWithOptions) {
+      throw new PollNotFoundError();
+    }
+    const pollOptionsIds = pollWithOptions.options.map((option) =>
+      option.id.toString()
+    );
+    if (!pollOptionsIds.includes(pollOptionId)) {
+      throw new InvalidPollOptionError();
+    }
+
     const userVote = await this.pollRepository.findUserPollVote(userId, pollId);
     if (userVote && userVote.pollOptionId.toString() === pollOptionId) {
-      throw new Error("User already voted on this option");
+      throw new DuplicateVoteError();
     }
     if (userVote) {
       await this.pollRepository.deleteUserPollVote(userId, pollId);
     }
 
     const vote = Vote.create({
-      pollOptionId: new UniqueEntityId(pollOptionId),
       userId: new UniqueEntityId(userId),
+      pollId: new UniqueEntityId(pollId),
+      pollOptionId: new UniqueEntityId(pollOptionId),
     });
 
     await this.pollRepository.createPollVote(vote);
